refactor classes in game.dart to separate files
