+++ doc/examples/particles/.gitignore
@@ -71,3 +71,7 @@
 !**/ios/**/default.pbxuser
 !**/ios/**/default.perspectivev3
 !/packages/flutter_tools/test/data/dart_dependencies_test/**/.packages
+
+.flutter-plugins-dependencies
+macos
+test
