+++ doc/examples/sprites/.gitignore
@@ -70,7 +70,5 @@
 !/packages/flutter_tools/test/data/dart_dependencies_test/**/.packages
 
 macos
-
 test
-
 .flutter-plugins-dependencies
