+++ example/.gitignore
@@ -71,3 +71,6 @@ build/
 !/packages/flutter_tools/test/data/dart_dependencies_test/**/.packages
 
 .flutter-plugins-dependencies
+
+macos
+test
