+++ lib/box2d/box2d_game.dart
@@ -2,7 +2,7 @@ import 'dart:ui';
 
 import 'package:flame/box2d/box2d_component.dart';
 import 'package:flame/components/component.dart';
-import 'package:flame/game.dart';
+import 'package:flame/game/base_game.dart';
 
 class Box2DGame extends BaseGame {
   final Box2DComponent box;
