+++ lib/components/composed_component.dart
@@ -2,7 +2,7 @@ import 'dart:ui';
 
 import 'package:flame/components/mixins/has_game_ref.dart';
 import 'package:flame/components/mixins/tapable.dart';
-import 'package:flame/game.dart';
+import 'package:flame/game/base_game.dart';
 import 'package:ordered_set/comparing.dart';
 import 'package:ordered_set/ordered_set.dart';
