+++ lib/game.dart
@@ -1,2 +1,3 @@
 // Keeping compatible with earlier versions of Flame
 export './game/game.dart';
+export './game/base_game.dart';
