+++ test/composed_component_test.dart
@@ -4,10 +4,10 @@ import 'package:flame/components/composed_component.dart';
 import 'package:flame/components/mixins/has_game_ref.dart';
 import 'package:flame/components/mixins/resizable.dart';
 import 'package:flame/components/mixins/tapable.dart';
+import 'package:flame/game/base_game.dart';
 import 'package:flutter/gestures.dart';
 import 'package:test/test.dart';
 
-import 'package:flame/game.dart';
 import 'package:flame/components/component.dart';
 
 class MyGame extends BaseGame {}
