+++ test/has_game_ref_test.dart
@@ -1,9 +1,9 @@
 import 'dart:ui';
 
 import 'package:flame/components/mixins/has_game_ref.dart';
+import 'package:flame/game/base_game.dart';
 import 'package:test/test.dart';
 
-import 'package:flame/game.dart';
 import 'package:flame/components/component.dart';
 
 class MyGame extends BaseGame {
