+++ test/resizable_test.dart
@@ -1,8 +1,8 @@
 import 'dart:ui';
 
+import 'package:flame/game/base_game.dart';
 import 'package:test/test.dart';
 
-import 'package:flame/game.dart';
 import 'package:flame/components/component.dart';
 import 'package:flame/components/mixins/resizable.dart';
